requirements.txt
streamlit==1.32.0
pypdf==4.1.0
langchain==0.1.12
langchain-groq==0.0.1
langchain-community==0.0.28
sentence-transformers==2.5.1
faiss-cpu==1.8.0
python-dotenv==1.0.1
reportlab==4.1.0